+++ server/bestbuy/pipelines.py
@@ -1,0 +1,23 @@
+# Define your item pipelines here
+#
+# Don't forget to add your pipeline to the ITEM_PIPELINES setting
+# See: https://docs.scrapy.org/en/latest/topics/item-pipeline.html
+
+
+# useful for handling different item types with a single interface
+from itemadapter import ItemAdapter
+from scrapy.exceptions import DropItem
+
+
+# Pipeline to remove duplicate products or reviews, from the docs: https://doc.scrapy.org/en/latest/topics/item-pipeline.html#duplicates-filter
+class DuplicatesPipeline:
+    def __init__(self):
+        self.ids_seen = set()
+
+    def process_item(self, item, spider):
+        adapter = ItemAdapter(item)
+        if adapter["id"] in self.ids_seen:
+            raise DropItem(f"Duplicate item found: {item!r}")
+        else:
+            self.ids_seen.add(adapter["id"])
+            return item
