+++ server/bestbuy/spiders/__init__.py
@@ -1,0 +1,4 @@
+# This package will contain the spiders of your Scrapy project
+#
+# Please refer to the documentation for information on how to create and manage
+# your spiders.
