--- server/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "best-buy-review-api",
-  "version": "1.0.0",
-  "main": "index.js",
-  "scripts": {
-    "start": "nodemon src/app.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "description": "",
-  "dependencies": {
-    "body-parser": "^1.20.2",
-    "cors": "^2.8.5",
-    "express": "^4.19.2",
-    "sqlite3": "^5.1.7",
-    "zeromq": "^6.0.0-beta.20"
-  },
-  "devDependencies": {
-    "nodemon": "^3.1.4"
-  }
-}
